# backends/java/src/main/java/com/example/multilang/dto/AuthResponse.java
package com.example.multilang.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class AuthResponse {
    private String token;
}